Feature/frontend projects page

+++ src/main/resources/static/js/app.js
@@ -1,13 +1,18 @@
+import "./components/alert-bar.js";
 import { showLoginInfoModal } from "./pages/login.js";
 import { toggleDropdownHeader } from "./pages/header.js";
 import { setupPasswordValidation } from "./pages/profile";
 import {
+  deleteCourseButton,
   updatePadding,
   showEmptyMessageCourses,
   deleteCourse,
 } from "./pages/courses.js";
+import {
+  showEmptyMessageProjects,
+  updateSectionProjectsHeight,
+} from "./pages/projects";
 // import "./pages/add-course.js"; // Load the add-course module
-import "./components/alert-bar.js";
 
 ///////////////////////////
 // Login Page
@@ -46,5 +51,22 @@ document.addEventListener("DOMContentLoaded", function (e) {
   showEmptyMessageCourses();
 
   // Deleting Course
-  deleteCourse();
+  if (deleteCourseButton) {
+    deleteCourse();
+  }
+});
+
+///////////////////////////
+// Projects Page
+
+// When the document is fully loaded
+document.addEventListener("DOMContentLoaded", function (e) {
+  //Start calculating header height
+  updateSectionProjectsHeight();
+
+  // Call updateSectionProjectsHeight when the window is resized
+  window.addEventListener("resize", updateSectionProjectsHeight);
+
+  // If no projects, show empty message
+  showEmptyMessageProjects();
 });
